+++ tests/tests.js
@@ -1,9 +1,9 @@
 'use strict';
 
-var xhr, requests;
-
 window.ParsleyConfig = $.extend( true, {}, window.ParsleyConfig, {
-  messages: {
+    // deactivate errors animation, causing travis test suite failing
+    animate: false
+  , messages: {
     type: {
       urlstrict: "urlstrict global override."
     }
@@ -416,6 +416,13 @@ var testSuite = function () {
           $( '#checkbox-mincheck2' ).attr( 'checked', 'checked' );
           expect( $( '#checkbox-mincheck1' ).parsley( 'validate' ) ).to.be( true );
         } )
+        it ( 'mincheck data-group', function () {
+          $( '#checkbox-mincheckgroup1' ).attr( 'checked', 'checked' );
+          expect( $( '#checkbox-mincheckgroup1' ).parsley( 'validate' ) ).to.be( false );
+          expect( getErrorMessage( '#checkbox-mincheckgroup1', 'mincheck') ).to.be( 'You must select at least 2 choices.' );
+          $( '#checkbox-mincheckgroup2' ).attr( 'checked', 'checked' );
+          expect( $( '#checkbox-mincheckgroup1' ).parsley( 'validate' ) ).to.be( true );
+        } )
         it ( 'maxcheck', function () {
           $( '#checkbox-maxcheck1' ).attr( 'checked', 'checked' );
           expect( $( '#checkbox-maxcheck1' ).parsley( 'validate' ) ).to.be( true );
@@ -463,46 +470,100 @@ var testSuite = function () {
         } )
 
         // not passing on phantomJS yet..
-        if ( !window.mochaPhantomJS ) {
-          describe ( 'Test ASYNC ajax calls results', function () {
-            var calls = [
-                { statusCode: 200, content: "true", expect: true }
-              , { statusCode: 404, content: "", expect: false }
-              , { statusCode: 200, content: "false", expect: false }
-              , { statusCode: 200, content: "1", expect: true }
-              , { statusCode: 200, content: "0", expect: false }
-              , { statusCode: 200, content: "{success: \"foobar\"}", expect: true }
-              ];
+        describe ( 'Test ASYNC ajax calls results', function () {
+          var calls = [
+              { statusCode: 'success', content: "true", expect: true }
+            , { statusCode: 'error', content: "", expect: false }
+            , { statusCode: 'success', content: "false", expect: false }
+            , { statusCode: 'success', content: "1", expect: true }
+            , { statusCode: 'success', content: "0", expect: false }
+            , { statusCode: 'success', content: "{\"success\": \"foo\"}", expect: true }
+            , { statusCode: 'success', content: "{\"error\": \"foobar\"}", expect: false }
+            , { statusCode: 'error', content: "{\"error\": \"foobarbaz\"}", expect: false }
+            , { statusCode: 'error', content: "{\"message\": \"foo\"}", expect: false }
+            ];
 
-            before( function () {
-              xhr = sinon.useFakeXMLHttpRequest();
-              requests = [];
-
-              xhr.onCreate = function ( xhr ) {
-                requests.push( xhr );
-              };
-            } )
-
-            it ( 'Test async ajax calls returns', function ( done ) {
-              $( '#remote2' ).val( 'foobarbaz' );
-
-              for ( var i = 0; i < calls.length; i++ ) {
-                $( '#remote2' ).parsley( 'validate' );
-                requests[ 0 ].respond([ calls[ i ].statusCode , {}, calls[ i ].content ]);
-                expect( requests.length ).to.be( 1 );
-                done();
-
-                expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ i ].expect)
-                expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ i ].expect );
-                expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ i ].expect );
-              }
-            } )
-
-            after( function () {
-              xhr.restore();
-            } );
+          it ( 'Test success true', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 0 ].statusCode , calls[ 0 ].content );
+            $( '#remote2' ).val( 'foo' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 0 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 0 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 0 ].expect );
+            done();
           } )
-        }
+          it ( 'Test error 404', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 1 ].statusCode , calls[ 1 ].content );
+            $( '#remote2' ).val( 'bar' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 1 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 1 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 1 ].expect );
+            done();
+          } )
+          it ( 'Test success false', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 2 ].statusCode , calls[ 2 ].content );
+            $( '#remote2' ).val( 'baz' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 2 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 2 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 2 ].expect );
+            done();
+          } )
+          it ( 'Test success 1', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 3 ].statusCode , calls[ 3 ].content );
+            $( '#remote2' ).val( 'foo' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 3 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 3 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 3 ].expect );
+            done();
+          } )
+          it ( 'Test success 0', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 4 ].statusCode , calls[ 4 ].content );
+            $( '#remote2' ).val( 'bar' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 4 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 4 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 4 ].expect );
+            done();
+          } )
+          it ( 'Test success with { success: "message" }', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 5 ].statusCode , calls[ 5 ].content );
+            $( '#remote2' ).val( 'baz' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 5 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 5 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 5 ].expect );
+            done();
+          } )
+          it ( 'Test success with { error: "message" } + display message ', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 6 ].statusCode , calls[ 6 ].content );
+            $( '#remote2' ).val( 'foo' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 6 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 6 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 6 ].expect );
+            expect( getErrorMessage( '#remote2', 'remote') ).to.be( 'foobar' );
+            done();
+          } )
+          it ( 'Test error 500 + error', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 7 ].statusCode , calls[ 7 ].content );
+            $( '#remote2' ).val( 'bar' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 7 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 7 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 7 ].expect );
+            expect( getErrorMessage( '#remote2', 'remote') ).to.be( 'foobarbaz' );
+            done();
+          } )
+
+          it ( 'Test error 500 + message', function ( done ) {
+            sinon.stub( $, "ajax" ).yieldsTo( calls[ 8 ].statusCode , calls[ 8 ].content );
+            $( '#remote2' ).val( 'baz' ).trigger( $.Event( 'change' ) );
+            expect( $( '#remote2' ).parsley( 'isFieldValid' ) ).to.be( calls[ 8 ].expect )
+            expect( $( '#remote2' ).hasClass( 'parsley-error' ) ).to.be( !calls[ 8 ].expect );
+            expect( $( '#remote2' ).hasClass( 'parsley-success' ) ).to.be( calls[ 8 ].expect );
+            expect( getErrorMessage( '#remote2', 'remote') ).to.be( 'foo' );
+            done();
+          } )
+
+          afterEach( function () {
+            $.ajax.restore();
+          } );
+        } )
 
       } )
     } )
@@ -534,6 +595,28 @@ var testSuite = function () {
         triggerSubmitValidation( '#requiredchanged3', '' );
         expect( getErrorMessage( '#requiredchanged3', 'type') ).to.be( 'custom email' );
         expect( getErrorMessage( '#requiredchanged3', 'required') ).to.be( 'custom required' );
+      } )
+      it ( 'Change error handler', function () {
+        $( '#errorsmanagement-form' ).parsley( {
+            successClass: 'parsley-great'
+          , errorClass: 'parsley-fail'
+          , errors: {
+            classHandler: function () {
+              return $( '#errorsmanagement-labelinfo' );
+            }
+            , container: function () {
+              return $( '#errorsmanagement-labelerror' );
+            }
+            , errorsWrapper: '<div></div>'
+            , errorElem: '<span></span>'
+          }
+        } );
+        $( '#errorsmanagement-email' ).val( 'foo' ).parsley( 'validate' );
+        expect( $( '#errorsmanagement-labelinfo' ).hasClass( 'parsley-fail' ) ).to.be( true );
+        expect( $( '#errorsmanagement-labelerror div.parsley-error-list span.type' ).length ).to.be( 1 );
+        $( '#errorsmanagement-email' ).val( 'foo@bar.baz' ).parsley( 'validate' );
+        expect( $( '#errorsmanagement-labelerror div' ).length ).to.be( 0 );
+        expect( $( '#errorsmanagement-labelinfo' ).hasClass( 'parsley-great' ) ).to.be( true );
       } )
     } )
 
@@ -603,7 +686,7 @@ var testSuite = function () {
         // The field has not changed, but since the field was reset, the call count should now be 2.
         expect( $( '#scenario-validation-after-field-reset' ).data( 'callCount' ) ).to.be( 2 );
       } )
-      it( 'Test always validate field', function () {
+      it ( 'Test always validate field', function () {
         $( '#alwaysValidate-form' ).parsley( { validateIfUnchanged: true, listeners: { onFieldError: function ( elem ) {
           if ( 'undefined' === typeof $( elem ).data( 'count' ) ) {
             $( elem ).data( 'count', 0 );
@@ -615,6 +698,26 @@ var testSuite = function () {
         $( '#alwaysValidate' ).parsley( 'validate' );
         $( '#alwaysValidate' ).parsley( 'validate' );
         expect( $( '#alwaysValidate' ).data( 'count' ) ).to.be( 3 );
+      } )
+      it ( 'Test data-trigger="change" on multiple inputs', function () {
+        $( '#checkbox-maxcheck1' ).parsley( 'reset' );
+        $( '#checkbox-maxcheck1' ).attr( 'checked', 'checked' );
+        $( '#checkbox-maxcheck2' ).attr( 'checked', 'checked' );
+        $( '#checkbox-maxcheck3' ).attr( 'checked', 'checked' ).trigger( $.Event( 'change' ) );
+        expect( $( 'ul#' + $( '#checkbox-maxcheck1' ).parsley( 'getHash' ) ).length ).to.be( 1 );
+        expect( getErrorMessage( '#checkbox-maxcheck1', 'maxcheck') ).to.be( 'You must select 2 choices or less.' );
+        $( '#checkbox-maxcheck3' ).attr( 'checked', null ).trigger( $.Event( 'change' ) );
+        expect( $( 'ul#' + $( '#checkbox-maxcheck1' ).parsley( 'getHash' ) ).length ).to.be( 0 );
+      } )
+      it ( 'Test change validation for checkboxes', function () {
+        $( '#checkbox-maxcheckchange1' ).attr( 'checked', 'checked' );
+        $( '#checkbox-maxcheckchange2' ).attr( 'checked', 'checked' );
+        $( '#checkbox-maxcheckchange3' ).attr( 'checked', 'checked' ).trigger( $.Event( 'change' ) );
+        expect( $( 'ul#' + $( '#checkbox-maxcheckchange1' ).parsley( 'getHash' ) ).length ).to.be( 0 );
+        expect( $( '#checkbox-maxcheckchange1' ).parsley( 'validate' ) ).to.be( false );
+        expect( $( 'ul#' + $( '#checkbox-maxcheckchange1' ).parsley( 'getHash' ) ).length ).to.be( 1 );
+        $( '#checkbox-maxcheckchange2' ).attr( 'checked', null ).trigger( $.Event( 'change' ) );
+        expect( $( 'ul#' + $( '#checkbox-maxcheckchange1' ).parsley( 'getHash' ) ).length ).to.be( 0 );
       } )
     } )
 
@@ -666,15 +769,28 @@ var testSuite = function () {
         expect( $( '#destroy-email' ).hasClass( 'parsley-error' ) ).to.be( false );
         expect( $( '#destroy-multiple' ).hasClass( 'parsley-error' ) ).to.be( false );
       } )
+      it ( 'test parsley(\'reset\')', function () {
+          $('#reset').parsley('validate');
+          expect( $('#reset-email').hasClass('parsley-error') ).to.be( true );
+          expect( $('#reset-textarea').hasClass('parsley-error') ).to.be( true );
+          $('#reset').parsley('reset');
+          expect( $('#reset-email').hasClass('parsley-error')).to.be( false );
+          expect( $('#reset-textarea').hasClass('parsley-error')).to.be( false );
+      })
       it ( 'test parsley dynamic add item', function () {
-        $( '#dynamic-form' ).append( '<input type="text" data-type="email" id="dynamic-email" data-trigger="change" value="foo" />' );
+        $( '#dynamic-form' ).append( '<input type="text" data-type="email" class="dynamic-email" data-trigger="change" value="foo" />' );
+        var ParsleyForm = $( '#dynamic-form' ).parsley();
+        expect( ParsleyForm.items.length ).to.be( 0 );
         expect( $( '#dynamic-form' ).parsley( 'validate' ) ).to.be( true );
-        $( '#dynamic-form' ).parsley( 'addItem', '#dynamic-email' );
+        $( '#dynamic-form' ).parsley( 'addItem', '.dynamic-email' );
+        expect( ParsleyForm.items.length ).to.be( 1 );
         expect( $( '#dynamic-form' ).parsley( 'validate' ) ).to.be( false );
-        $( '#dynamic-email' ).val( 'foo@bar.baz' );
+        $( '.dynamic-email' ).val( 'foo@bar.baz' );
         expect( $( '#dynamic-form' ).parsley( 'validate' ) ).to.be( true );
-        $( '#dynamic-email' ).val( 'foo' );
-        $( '#dynamic-form' ).parsley( 'removeItem', '#dynamic-email' );
+        $( '.dynamic-email' ).val( 'foo' );
+        $( '#dynamic-form' ).parsley( 'removeItem', '.dynamic-email' );
+        expect( ParsleyForm.items.length ).to.be( 0 );
+        expect( $( '.dynamic-email' ).hasClass( 'parsley-validated' ) ).to.be( false );
         expect( $( '#dynamic-form' ).parsley( 'validate' ) ).to.be( true );
       } )
       it ( 'test adding constraint on the fly', function () {